+++ doc/changes.rst
@@ -6,8 +6,10 @@ desispec Change Log
 -------------------
 
 * Adds `desi_pipe go` for production running (PR `#666`_).
+* Increase job maxtime for edison realtime queue (PR `#667`_).
 
 .. _`#666`: https://github.com/desihub/desispec/pull/666
+.. _`#667`: https://github.com/desihub/desispec/pull/667
 
 0.22.1 (2018-07-18)
 -------------------